pyproject.toml
[project]
name = "eosergiobot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.49.0",
    "mypy",
    "py-hot-reload>=1.0.6",
    "python-dotenv>=1.2.1",
    "ruff",
    "telebot>=0.0.5",
]
